Moved addOn management inside ConfigCommandAddOn
